pyproject.toml
[project]
name = "ai-agent-learning"
version = "0.1.0"
description = "AI Agent 开发学习仓库"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ipython>=8.0.0",
]
